backend/requirements.txt
# AWS SDK
boto3==1.34.69
botocore>=1.34.69,<1.35.0

# Database
psycopg2-binary==2.9.1
SQLAlchemy==2.0.27
alembic==1.13.1
sqlalchemy-utils==0.41.1

# JWT handling
PyJWT==2.1.0
cryptography==41.0.1

# HTTP handling
requests==2.31.0

# Utilities
python-dotenv==0.19.0
pytest==8.1.1
pytest-cov==4.1.0
localstack-client==2.5
werkzeug==2.0.1

# Flask
flask==2.0.1
flask-cors==3.0.10

# OpenAI
openai==1.3.0 